pyproject.toml
[project]
name = "bot-master"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "textual>=3.0",
    "pyyaml>=6.0",
]

[project.scripts]
bot-master-daemon = "bot_master.daemon:main"
bot-master = "bot_master.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
